+++ lib/contentstack/region.rb
@@ -2,5 +2,13 @@ module Contentstack
   class Region
     EU='eu'
     US='us'
+    AZURE_NA='azure-na'
+    AZURE_EU='azure-eu'
+  end
+
+  class Host
+    PROTOCOL='https://'
+    DEFAULT_HOST='cdn.contentstack.io'
+    HOST='contentstack.com'
   end
 end